+++ sql-to-dbsp-compiler/SQL-compiler/src/main/java/org/dbsp/sqlCompiler/compiler/frontend/ExpressionCompiler.java
@@ -537,6 +537,14 @@ public class ExpressionCompiler extends RexVisitorImpl<DBSPExpression> implement
             ops.set(argument, arg.cast(new DBSPTypeDouble(arg.getType().getNode(), arg.getType().mayBeNull)));
     }
 
+    void ensureInteger(List<DBSPExpression> ops, int argument, int length) {
+        DBSPExpression arg = ops.get(argument);
+        DBSPTypeInteger expected = new DBSPTypeInteger(arg.getType().getNode(), length, true, arg.getType().mayBeNull);
+
+        if (!arg.getType().sameType(expected))
+            ops.set(argument, arg.cast(expected));
+    }
+
     void nullLiteralToNullArray(List<DBSPExpression> ops, int arg) {
         if (ops.get(arg).is(DBSPNullLiteral.class)) {
             ops.set(arg, new DBSPTypeVec(new DBSPTypeNull(CalciteObject.EMPTY), true).nullValue());
@@ -979,6 +987,11 @@ public class ExpressionCompiler extends RexVisitorImpl<DBSPExpression> implement
                         return makeBinaryExpressions(node, type, DBSPOpcode.CONCAT, ops);
                     case "array":
                         return this.compileFunction(call, node, type, ops, 0);
+                    case "sequence":
+                        for (int i = 0; i < ops.size(); i++)
+                            this.ensureInteger(ops, i, 32);
+
+                        return this.compileFunction(call, node, type, ops, 2);
                 }
                 return this.compileUDF(node, call, type, ops);
             }