+++ sql-to-dbsp-compiler/SQL-compiler/src/main/java/org/dbsp/sqlCompiler/compiler/frontend/calciteCompiler/CustomFunctions.java
@@ -19,6 +19,7 @@ import org.apache.calcite.sql.type.SqlOperandTypeInference;
 import org.apache.calcite.sql.type.SqlSingleOperandTypeChecker;
 import org.apache.calcite.sql.type.SqlTypeFamily;
 import org.apache.calcite.sql.type.SqlTypeName;
+import org.apache.calcite.sql.type.SqlTypeTransforms;
 import org.dbsp.sqlCompiler.compiler.errors.CompilationError;
 import org.dbsp.sqlCompiler.compiler.frontend.CalciteObject;
 import org.dbsp.util.Linq;
@@ -41,6 +42,7 @@ public class CustomFunctions {
     public CustomFunctions() {
         this.initial.add(new RlikeFunction());
         this.initial.add(new WriteLogFunction());
+        this.initial.add(new SequenceFunction());
         this.udf = new HashMap<>();
     }
 
@@ -49,7 +51,7 @@ public class CustomFunctions {
         public RlikeFunction() {
             super("RLIKE",
                     SqlKind.RLIKE,
-                    ReturnTypes.BOOLEAN,
+                    ReturnTypes.BOOLEAN_NULLABLE,
                     null,
                     OperandTypes.STRING_STRING,
                     SqlFunctionCategory.STRING);
@@ -73,6 +75,28 @@ public class CustomFunctions {
                     ARG1,
                     null,
                     family(SqlTypeFamily.CHARACTER, SqlTypeFamily.ANY),
+                    SqlFunctionCategory.USER_DEFINED_FUNCTION);
+        }
+
+        @Override
+        public boolean isDeterministic() {
+            return false;
+        }
+    }
+
+    /**
+     * SEQUENCE(start, end) returns an array of integers from start to end (inclusive).
+     * The array is empty if start > end.
+     */
+    public static class SequenceFunction extends SqlFunction {
+        public SequenceFunction() {
+            super("SEQUENCE",
+                    SqlKind.OTHER_FUNCTION,
+                    ReturnTypes.INTEGER
+                            .andThen(SqlTypeTransforms.TO_ARRAY)
+                            .andThen(SqlTypeTransforms.TO_NULLABLE),
+                    null,
+                    family(SqlTypeFamily.INTEGER, SqlTypeFamily.INTEGER),
                     SqlFunctionCategory.USER_DEFINED_FUNCTION);
         }