+++ CHANGELOG.md
@@ -23,9 +23,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - dbsp: Major refactoring of the dbsp crate to improve
   to improve compilation speed of SQL programs
   ([#1516](https://github.com/feldera/feldera/pull/1516))
+- SQL: Fix `RLIKE` to handle `NULL`s correctly ([#1555](https://github.com/feldera/feldera/pull/1555))
 
 ### Added
 
+- SQL: Support for `SEQUENCE` function ([#1555](https://github.com/feldera/feldera/pull/1555))
 - SQL: Support for `GROUPING`, `GROUPING_ID`, and `GROUPING SETS`
   ([#1527](https://github.com/feldera/feldera/pull/1527))
 - SQL: Support for aggregation functions: `STDDEV`,